reggie-app-runner/pyproject.toml
[build-system]
requires = [
    "uv_build>=0.8.23,<0.9.0",
]
build-backend = "uv_build"

[project]
name = "reggie-app-runner"
version = "0.0.1+g4b7f333"
requires-python = ">=3.10"
dependencies = [
    "reggie-core @ file://${PROJECT_ROOT}/../reggie-core",
    "dynaconf",
    "pyaml",
    "sh",
    "requirements-parser",
    "giturlparse",
]

[tool.uv.sources.reggie-core]
workspace = true
